Split geocoder/carrier/timezone into their own packages
Mirrors upstream's module layout (separate carrier/geocoder modules plus a
shared internal/prefixmapper) and finishes the 2.0 cleanup (#4 in
docs/2.0-plan.md). The flat root functions move to dedicated packages:

  phonenumbers.GetGeocodingForNumber          -> geocoding.ForNumber
  phonenumbers.GetCarrierForNumber            -> carrier.ForNumber
  phonenumbers.GetCarrierWithPrefixForNumber  -> carrier.WithPrefixForNumber
  phonenumbers.GetSafeCarrierDisplayName...   -> carrier.SafeDisplayName
  phonenumbers.GetTimezonesForNumber          -> timezone.ForNumber
  phonenumbers.GetTimezonesForPrefix          -> timezone.ForPrefix
  phonenumbers.UNKNOWN_TIMEZONE               -> timezone.Unknown

The shared prefix lookup becomes internal/prefixmapper.Mapper: a value that
owns its own mutex and per-language cache, decoupled from the root package
(it operates on the E164 string rather than a *PhoneNumber). carrier and
geocoding each construct one over their own embedded data.

Each package owns its embed, so the data relocated under it (carrier/data,
geocoding/data, timezone/data); go:embed can't reach a sibling directory, the
same reason metadata/data moved earlier. cmd/buildmetadata writes the new
paths. The geocoding/carrier/timezone tests move to the new packages.

This delivers a real size win, not just upstream parity. PhoneNumberOffline-
Geocoder pulls in golang.org/x/text for its locale fallback, whose display
tables have package-init side effects the linker can't dead-code-eliminate.
Because the geocoder lived in root, every consumer linked in ~2.6 MB of CLDR
data even when only calling Parse/Format. A core-only binary now drops from
9.33 MB to 6.74 MB; x/text and the geocoding data link in only when the
geocoding package is imported. The root package no longer depends on x/text.

Also removes stale .gitignore entries (bare "carrier" and "geocoding", once
build-output names) that now matched the new package directories.
