+++ .gitignore
@@ -5,8 +5,6 @@
 deploy
 fabfile.py
 fabfile.pyc
-carrier
-geocoding
 functions/*
 dist/
 _build
