+++ carrier/data/ko.txt.gz
