+++ geocoding/geocoding_test.go
@@ -1,0 +1,39 @@
+package geocoding
+
+import (
+	"testing"
+
+	"github.com/nyaruka/phonenumbers/v2"
+)
+
+func TestForNumber(t *testing.T) {
+	tests := []struct {
+		num      string
+		lang     string
+		expected string
+	}{
+		{num: "+8613702032331", lang: "en", expected: "Tianjin"},
+		{num: "+8613702032331", lang: "zh", expected: "天津市"},
+		{num: "+863197785050", lang: "zh", expected: "河北省邢台市"},
+		{num: "+8613323241342", lang: "en", expected: "Baoding, Hebei"},
+		{num: "+917999499543", lang: "en", expected: "Ahmedabad Local, Gujarat"},
+		{num: "+17047181840", lang: "en", expected: "North Carolina"},
+		{num: "+12542462158", lang: "en", expected: "Texas"},
+		{num: "+16193165996", lang: "en", expected: "California"},
+		{num: "+12067799191", lang: "en", expected: "Washington State"},
+		{num: "+447825602614", lang: "en", expected: "United Kingdom"},
+	}
+	for _, test := range tests {
+		number, err := phonenumbers.Parse(test.num, "ZZ")
+		if err != nil {
+			t.Errorf("Failed to parse number %s: %s", test.num, err)
+		}
+		geocoding, err := ForNumber(number, test.lang)
+		if err != nil {
+			t.Errorf("Failed to getGeocoding for the number %s: %s", test.num, err)
+		}
+		if test.expected != geocoding {
+			t.Errorf("Expected '%s', got '%s' for '%s'", test.expected, geocoding, test.num)
+		}
+	}
+}
