+++ phonenumberutil_adhoc_test.go
@@ -20,7 +20,6 @@ package phonenumbers
 
 import (
 	"reflect"
-	"sync"
 	"testing"
 
 	"github.com/nyaruka/phonenumbers/v2/metadata"
@@ -96,253 +95,6 @@ func newPhoneNumber(cc int, natNum uint64) *PhoneNumber {
 	return p
 }
 
-type timeZonesTestCases struct {
-	num               string
-	expectedTimeZones []string
-}
-
-func TestGetTimeZonesForPrefix(t *testing.T) {
-	tests := []timeZonesTestCases{
-		{
-			num:               "+442073238299",
-			expectedTimeZones: []string{"Europe/London"},
-		},
-		{
-			num:               "+61491570156",
-			expectedTimeZones: []string{"Australia/Sydney"},
-		},
-		{
-			num:               "+61255501234",
-			expectedTimeZones: []string{"Australia/Sydney"},
-		},
-		{
-			num:               "+12067798181",
-			expectedTimeZones: []string{"America/Los_Angeles"},
-		},
-		{
-			num:               "+390399123456",
-			expectedTimeZones: []string{"Europe/Rome"},
-		},
-		{
-			num:               "+541151123456",
-			expectedTimeZones: []string{"America/Buenos_Aires"},
-		},
-		{
-			num:               "+15167706076",
-			expectedTimeZones: []string{"America/New_York"},
-		},
-		{
-			num:               "+917999999543",
-			expectedTimeZones: []string{"Asia/Calcutta"},
-		},
-		{
-			num:               "+540111561234567",
-			expectedTimeZones: []string{"America/Buenos_Aires"},
-		},
-		{
-			num:               "+18504320800",
-			expectedTimeZones: []string{"America/Chicago"},
-		},
-		{
-			num:               "+14079395277",
-			expectedTimeZones: []string{"America/New_York"},
-		},
-		{
-			num:               "+18508632167",
-			expectedTimeZones: []string{"America/Chicago"},
-		},
-		{
-			num:               "+40213158207",
-			expectedTimeZones: []string{"Europe/Bucharest"},
-		},
-		// UTC +5:45
-		{
-			num:               "+97714240520",
-			expectedTimeZones: []string{"Asia/Katmandu"},
-		},
-		// UTC -3:30
-		{
-			num:               "+17097264534",
-			expectedTimeZones: []string{"America/St_Johns"},
-		},
-		{
-			num:               "0000000000",
-			expectedTimeZones: []string{"Etc/Unknown"},
-		},
-		{
-			num:               "+31112",
-			expectedTimeZones: []string{"Europe/Amsterdam"},
-		},
-		{
-			num:               "+6837000",
-			expectedTimeZones: []string{"Pacific/Niue"},
-		},
-		{
-			num: "+1911",
-			expectedTimeZones: []string{
-				"America/Adak", "America/Anchorage", "America/Anguilla", "America/Antigua",
-				"America/Barbados", "America/Boise", "America/Cayman", "America/Chicago",
-				"America/Denver", "America/Dominica", "America/Edmonton", "America/Fort_Nelson",
-				"America/Grand_Turk", "America/Grenada", "America/Halifax", "America/Jamaica",
-				"America/Juneau", "America/Los_Angeles", "America/Lower_Princes", "America/Montserrat",
-				"America/Nassau", "America/New_York", "America/North_Dakota/Center",
-				"America/Phoenix", "America/Port_of_Spain", "America/Puerto_Rico",
-				"America/Regina", "America/Santo_Domingo", "America/St_Johns", "America/St_Kitts",
-				"America/St_Lucia", "America/St_Thomas", "America/St_Vincent", "America/Toronto",
-				"America/Tortola", "America/Vancouver", "America/Winnipeg", "Atlantic/Bermuda",
-				"Pacific/Guam", "Pacific/Honolulu", "Pacific/Pago_Pago", "Pacific/Saipan",
-			},
-		},
-	}
-
-	for _, test := range tests {
-		timeZones, err := GetTimezonesForPrefix(test.num)
-		if err != nil {
-			t.Errorf("Failed to getTimezone for the number %s: %s", test.num, err)
-		}
-
-		if len(timeZones) == 0 {
-			t.Errorf("Expected at least 1 timezone.")
-		}
-
-		if !reflect.DeepEqual(timeZones, test.expectedTimeZones) {
-			t.Errorf("Expected '%v', got '%v' for '%s'", test.expectedTimeZones, timeZones, test.num)
-		}
-
-		num, err := Parse(test.num, "")
-		if err != nil {
-			continue
-		}
-
-		timeZones, err = GetTimezonesForNumber(num)
-		if err != nil {
-			t.Errorf("Failed to getTimezone for the number %s: %s", num, err)
-		}
-
-		if len(timeZones) == 0 {
-			t.Errorf("Expected at least 1 timezone.")
-		}
-
-		if !reflect.DeepEqual(timeZones, test.expectedTimeZones) {
-			t.Errorf("Expected '%v', got '%v' for '%s'", test.expectedTimeZones, timeZones, num)
-		}
-	}
-}
-
-func TestGetCarrierForNumber(t *testing.T) {
-	tests := []struct {
-		num      string
-		lang     string
-		expected string
-	}{
-		{num: "+8613702032331", lang: "en", expected: "China Mobile"},
-		{num: "+8613702032331", lang: "zh", expected: "中国移动"},
-		{num: "+6281377468527", lang: "en", expected: "Telkomsel"},
-		{num: "+8613323241342", lang: "en", expected: "China Telecom"},
-		{num: "+61491570156", lang: "en", expected: "Telstra"},
-		{num: "+917999999543", lang: "en", expected: "Reliance Jio"},
-		{num: "+593992218722", lang: "en", expected: "Claro"},
-	}
-	for _, test := range tests {
-		number, err := Parse(test.num, "ZZ")
-		if err != nil {
-			t.Errorf("Failed to parse number %s: %s", test.num, err)
-		}
-		carrier, err := GetCarrierForNumber(number, test.lang)
-		if err != nil {
-			t.Errorf("Failed to getCarrier for the number %s: %s", test.num, err)
-		}
-		if test.expected != carrier {
-			t.Errorf("Expected '%s', got '%s' for '%s'", test.expected, carrier, test.num)
-		}
-	}
-}
-
-func TestGetCarrierWithPrefixForNumber(t *testing.T) {
-	tests := []struct {
-		num             string
-		lang            string
-		expectedCarrier string
-		expectedPrefix  int
-	}{
-		{num: "+8613702032331", lang: "en", expectedCarrier: "China Mobile", expectedPrefix: 86137},
-		{num: "+8613702032331", lang: "zh", expectedCarrier: "中国移动", expectedPrefix: 86137},
-		{num: "+6281377468527", lang: "en", expectedCarrier: "Telkomsel", expectedPrefix: 62813},
-		{num: "+8613323241342", lang: "en", expectedCarrier: "China Telecom", expectedPrefix: 86133},
-		{num: "+61491570156", lang: "en", expectedCarrier: "Telstra", expectedPrefix: 6149},
-		{num: "+917999999543", lang: "en", expectedCarrier: "Reliance Jio", expectedPrefix: 917999},
-		{num: "+593992218722", lang: "en", expectedCarrier: "Claro", expectedPrefix: 5939922},
-		{num: "+201987654321", lang: "en", expectedCarrier: "", expectedPrefix: 0},
-		{num: "+201987654321", lang: "notFound", expectedCarrier: "", expectedPrefix: 0},
-	}
-	for _, test := range tests {
-		number, err := Parse(test.num, "ZZ")
-		if err != nil {
-			t.Errorf("Failed to parse number %s: %s", test.num, err)
-		}
-		carrier, prefix, err := GetCarrierWithPrefixForNumber(number, test.lang)
-		if err != nil {
-			t.Errorf("Failed to getCarrierWithPrefix for the number %s: %s", test.num, err)
-		}
-		if test.expectedCarrier != carrier {
-			t.Errorf("Expected '%s', got '%s' for '%s'", test.expectedCarrier, carrier, test.num)
-		}
-		if test.expectedPrefix != prefix {
-			t.Errorf("Expected '%d', got '%d' for '%s'", test.expectedPrefix, prefix, test.num)
-		}
-	}
-}
-
-func TestGetCarrierWithPrefixForNumberWithConcurrency(t *testing.T) {
-	number, _ := Parse("+8613702032331", "ZZ")
-
-	wg := sync.WaitGroup{}
-
-	for i := 0; i < 10; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			_, _, err := GetCarrierWithPrefixForNumber(number, "en")
-			if err != nil {
-				t.Errorf("Failed to getCarrierWithPrefix for the number %s: %s", "+8613702032331", err)
-			}
-		}()
-	}
-
-	wg.Wait()
-}
-
-func TestGetGeocodingForNumber(t *testing.T) {
-	tests := []struct {
-		num      string
-		lang     string
-		expected string
-	}{
-		{num: "+8613702032331", lang: "en", expected: "Tianjin"},
-		{num: "+8613702032331", lang: "zh", expected: "天津市"},
-		{num: "+863197785050", lang: "zh", expected: "河北省邢台市"},
-		{num: "+8613323241342", lang: "en", expected: "Baoding, Hebei"},
-		{num: "+917999499543", lang: "en", expected: "Ahmedabad Local, Gujarat"},
-		{num: "+17047181840", lang: "en", expected: "North Carolina"},
-		{num: "+12542462158", lang: "en", expected: "Texas"},
-		{num: "+16193165996", lang: "en", expected: "California"},
-		{num: "+12067799191", lang: "en", expected: "Washington State"},
-		{num: "+447825602614", lang: "en", expected: "United Kingdom"},
-	}
-	for _, test := range tests {
-		number, err := Parse(test.num, "ZZ")
-		if err != nil {
-			t.Errorf("Failed to parse number %s: %s", test.num, err)
-		}
-		geocoding, err := GetGeocodingForNumber(number, test.lang)
-		if err != nil {
-			t.Errorf("Failed to getGeocoding for the number %s: %s", test.num, err)
-		}
-		if test.expected != geocoding {
-			t.Errorf("Expected '%s', got '%s' for '%s'", test.expected, geocoding, test.num)
-		}
-	}
-}
 func TestMaybeStripExtension(t *testing.T) {
 	var tests = []struct {
 		input     string
@@ -525,30 +277,6 @@ func TestRegexCacheStrict(t *testing.T) {
 	}
 }
 
-func TestGetSafeCarrierDisplayNameForNumber(t *testing.T) {
-	tests := []struct {
-		num      string
-		lang     string
-		expected string
-	}{
-		{num: "+447387654321", lang: "en", expected: ""},
-		{num: "+244917654321", lang: "en", expected: "Movicel"},
-	}
-	for _, test := range tests {
-		number, err := Parse(test.num, "ZZ")
-		if err != nil {
-			t.Errorf("Failed to parse number %s: %s", test.num, err)
-		}
-		carrier, err := GetSafeCarrierDisplayNameForNumber(number, test.lang)
-		if err != nil {
-			t.Errorf("Failed to getSafeCarrierDisplayNameForNumber for the number %s: %s", test.num, err)
-		}
-		if test.expected != carrier {
-			t.Errorf("Expected '%s', got '%s' for '%s'", test.expected, carrier, test.num)
-		}
-	}
-}
-
 func TestFormattingRuleHasFirstGroupOnly(t *testing.T) {
 	// Verify the fix: formattingRuleHasFirstGroupOnly should do a full match
 	assert.True(t, formattingRuleHasFirstGroupOnly("$1"))
@@ -562,16 +290,5 @@ func TestFormattingRuleHasFirstGroupOnly(t *testing.T) {
 func BenchmarkLoadMetadata(b *testing.B) {
 	for i := 0; i <= b.N; i++ {
 		_, _ = metadata.Load()
-	}
-}
-
-func BenchmarkGetCarrierWithPrefixForNumber(b *testing.B) {
-	number, _ := Parse("+8613702032331", "ZZ")
-
-	for n := 0; n < b.N; n++ {
-		_, _, err := GetCarrierWithPrefixForNumber(number, "en")
-		if err != nil {
-			b.Errorf("Failed to getCarrierWithPrefix for the number %s: %s", "+8613702032331", err)
-		}
 	}
 }
