+++ timezone/timezone.go
@@ -1,20 +1,35 @@
 // Port of geocoder/PhoneNumberToTimeZonesMapper.java from google/libphonenumber.
 // Functions are kept in upstream source order to ease syncing.
-package phonenumbers
+package timezone
 
 import (
+	_ "embed"
 	"fmt"
 	"strconv"
 	"strings"
+	"sync"
 
+	"github.com/nyaruka/phonenumbers/v2"
 	"github.com/nyaruka/phonenumbers/v2/internal/serialize"
 )
 
-// GetTimezonesForPrefix returns a slice of Timezones corresponding to the number passed
+//go:embed data/prefix_to_timezone.xml.gz
+var timezoneData []byte
+
+// Unknown is the value returned for a prefix that maps to no timezone. This is
+// defined by ICU as the unknown time zone.
+const Unknown = "Etc/Unknown"
+
+var (
+	timezoneOnce sync.Once
+	timezoneMap  *serialize.IntStringArrayMap
+)
+
+// ForPrefix returns a slice of Timezones corresponding to the number passed
 // or error when it is impossible to convert the string to int
 // The algorithm tries to match the timezones starting from the maximum
 // number of phone number digits and decreasing until it finds one or reaches 0
-func GetTimezonesForPrefix(number string) ([]string, error) {
+func ForPrefix(number string) ([]string, error) {
 	var err error
 	timezoneOnce.Do(func() {
 		timezoneMap, err = serialize.LoadIntArrayMap(timezoneData)
@@ -39,12 +54,12 @@ func GetTimezonesForPrefix(number string) ([]string, error) {
 			return tzs, nil
 		}
 	}
-	return []string{UNKNOWN_TIMEZONE}, nil
+	return []string{Unknown}, nil
 }
 
-// GetTimezonesForNumber returns the names of timezones which we believe maps to the
+// ForNumber returns the names of timezones which we believe maps to the
 // passed in number.
-func GetTimezonesForNumber(number *PhoneNumber) ([]string, error) {
-	e164 := Format(number, E164)
-	return GetTimezonesForPrefix(e164)
+func ForNumber(number *phonenumbers.PhoneNumber) ([]string, error) {
+	e164 := phonenumbers.Format(number, phonenumbers.E164)
+	return ForPrefix(e164)
 }
